Upgrade to net9.0 and revive project
It's been a busy year, but it's time to wrap this up.

In trying to be smart with 3e2168f I actually introduced a bug where
the Resolver would add +1 distance to function body statements, because
it would visit the body _as a block_, and blocks get their own scope. It
could visit the body as a list and the bug goes away, but the experience
of chasing this down made me realize: maintaining AST types that are
different from the book just opens the door to bugs more than it offers
some ergonomic advantage. So that's been reverted. In that spirit, I
also removed the error type base class and moved the core error message
logic back into the Lox class.

I also upped the analysis mode on the two projects here to
"recommended," and cleaned up the ensuing errors.

Last, I did some light housekeeping -- more comments, clearer
folder/identifier names, simpler namespaces.

Now it's back to chapter 12 and implementing classes.