Add class constructors
The me-diverging-from-the-book chicken has come home to roost yet again.
If user code omits a variable stmt initializer or a return value, I'd
made it a concern of the Parser to fill in that omission with a literal
expr that stores the nil instance. But that just means we can't detect
downstream whether the user _actually_ omitted the thing, or just
specified an explicit nil (because the result is the same either way).
That wasn't really a problem until now, when we only want to allow a
return stmt in a class constructor if the user fully omits the value.
So, I retyped those properties as nullable, parse them as null if
they're not present, and check whether they're null before
resolving/interpreting. I continue to think it's a good idea, however,
to keep the value of a literal expr non-nullable, because it means you
can assume environment values are non-null too.

Back to the core change here: if your class definition has an "init"
method, it will run each time you create a new instance. Init will
always return the "this" to which it belongs (you can call it ad hoc if
you so choose). Finally, you can return from init, but you may not
specify an explicit return value.