Call superclass methods
And with that, Lox is feature complete*. Left to do:
- *It might be fun to define a native list class though.
- When you VisitClassStmt in the Interpreter, the book has you define
the class name, then build the LoxClass, then assign it (under the
premise that it allows methods to reference the class). That makes sense
for resolving b/c you resolve methods as you go, but I don't think it's
necessary here (you don't _evaluate_ the methods as you go). I'll write
an issue on the book's repo about this.
- Add/clean up docstrings.
- Write some test scripts (maybe as proper unit tests?), and link to
them when I...
- Add more detail to the README.